Initial work on ordering hands
A few things in this commit.
1. Hand-to-index mapper, like for ranks.
2. Start on tiebreaking, the more complex aspect of
   hand comparison. Mostly not done yet.
3. Stub PartialEq impl.

Some tricky lessons-learned:
Struct-like enum variants! I read these are
"experimental." Make sure to always use the
curlies for them, which has been an unintuive
gotcha.
Also, when binding the contents of struct-like
enum variants to a variable in a pattern, the
syntax makes no sense to me; I discovered it by
trying random things. It's

struct_field_name: local_variable_name

I would think there would be an @ in there
somewhere, or that types would be involved if
there's a :. But nope. Maybe I just don't
properly understand this yet.