Compiler version upgrade, and start of handmaking
Big commit! Sorry!

The compiler version is upgraded here.
-Show becomes Debug and is sometimes derive'd. Sweet!
-Ord/Eq can now be impl'ed directly. Don't think this was possible
before.

Also, started working on making hands out of cards. The skeleton is
there, but I need to fill in a lot of the meat. The basic algorithm is
to try to form all hands in descending order, so that each hand-attempt
function can assume that it is not being passed cards that form a better
hand than the one it's attempting to make.

Other stuff:
-Make Hand not have "struct variants" or whatever, but instead make
explicit structs for each type of Hand and have each of them just
contain that struct. This lets me refer to a particular type of hand.
Reddit sez what I want is called a "refinement type" and is a
long-standing language feature request.
-pub everything.
-copy_all function, for copying all elements out of one vector into
another. Maybe this already exists somewhere. There is Vec::push_all,
but this clones, whereas I want to copy. (These are different? I think?)
There is Vec::append, but that mutates the source vector, which I don't
want. It also moves the members, and I'm not sure if that's what I want.