Test method that makes hands ordered by rank and compares them
This is getting to be a common pattern in my test code, and
it's getting me to learn about closures and iterators and
stuff.

This goes over all ranks in order and makes at most one
hand per rank, and assumes that the rank should represent
the strength of the hand. It builds a list and makes sure
everything in it has the right order with respect to
everything else.

Note that this "at most one hand per rank" thing does not
seem ideal - I think what I'd really like is an arbitary
adapter from iterator of ranks to iterator of hands, that's
free to produce an arbitrary number of elements per rank,
and reorder, and do whatever else. (If reordering is
omitted, I guess this is covered by flat_map. Actually, I
may not actually care about reordering, since ordering by
rank is the whole trick here.) The point is, in this case
I'm fine with filter_map, but it seems like there may be an
arbitrary "adapter" type out there but I don't know about
it.