Convert existing tests to cmp_by_rank
Slightly fewer lines of code, somewhat fewer characters, a
fair bit less duplication of logic, and a bit of memory
bloat (the vector is now unprecisely sized to 13 as opposed
to the actual number of hands produced, which it doesn't
know ahead of time).

And some lessons learned.

Seems like a good tradeoff!