Smoke test of trips comparisons
Also improve what the tests print out to actually be
useful. I've basically written an "assert_cmp" macro, but
not actually.

The reason I had to do this, besides it being a good idea,
is that this test kept failing because I was making it
expect wrong orders by screwing up the list. It passes now,
but at this point I have no confidence that there is not
actually a bug which is counteracting some remaining mis-
ordering in my manually constructed list.

I'll write some more "try all the possibilities" tests
next...