Start hand-making tests
-Add convenience card "factory" method. Or should it be a
macro? Anyway, it's meant to be shorthand instead of having
to type something like "Hand{Suit: Spades, Rank: Ace}" for
every single hand I want to specify. There's gonna be a
LOT.

-Fix an actual bug! Caught by the test. Turns out an assert
I had was doing the wrong thing. So it's still only a "test
bug."

-Box autodereference! &box is the same as &*box, which
borrows the contents of the box.