Logic for making everyone's hands
To do this, I had to make Hand (and everything in it) be
Copy (which also needs Clone).

I also exposed the push_all() function from poker_hands,
since it's so useful.

There is actually a similar function on Vector: extend().
But that doesn't seem to work with slices since they
iterate over borrows of their elements. That's ok for me
because the elements are Copy anyway, but extend() doesn't
seem to know or care.

I think the Cloned iterator struct does what I want (seems
to adapt an iterator over Clone refs into an iterator over
the underlying type) but it's unstable.